Chapter 4/requirements.txt
openai==1.3.7
google-cloud-language==2.13.3
google-cloud-vision==3.7.2
google-cloud-translate==3.11.3
langchain==0.2.0
langchain-community==0.2.0
langchain-anthropic
langchain-cohere
cohere==5.5.0
torch
transformers==4.40.2
accelerate==0.30.1
huggingface-hub
